requirements.txt
requests
stups-zign
